+++ coc/static/update_static.py
@@ -432,6 +432,7 @@ class StaticUpdater:
                     "required_townhall": level_data.get("TownHallLevel"),
                     "hitpoints": level_data.get("Hitpoints", 0),
                     "dps": level_data.get("DPS", 0) or level_data.get("Damage", 0),
+                    "strength_weight": level_data.get("StrengthWeight", 0)
                 }
 
                 if "AltBuildResource" in level_data:
@@ -470,6 +471,7 @@ class StaticUpdater:
                                 "info": weapon_data.get("InfoTID"),
                             },
                             "upgrade_resource": self._parse_resource(resource=building_data.get("BuildResource")),
+                            "strength_weight": level_data.get("StrengthWeight", 0),
                             "levels": []
                         }
                         for weapon_level, weapon_level_data in weapon_data.items():
@@ -632,6 +634,7 @@ class StaticUpdater:
                 "attack_speed": troop_data.get("AttackSpeed", 0),
                 "attack_range": troop_data.get("AttackRange", 0),
                 "housing_space": troop_data.get("HousingSpace"),
+
                 "village": "home" if not village_type else "builderBase",
             }
             is_super_troop = troop_data.get("EnabledBySuperLicence", False)
@@ -677,6 +680,8 @@ class StaticUpdater:
                     "upgrade_cost": level_data.get("UpgradeCost", 0),
                     "required_lab_level": required_lab_level,
                     "required_townhall": required_townhall,
+
+                    "strength_weight": level_data.get("StrengthWeight", 0),
                 }
                 hold_data["levels"].append(new_level_data)
 
@@ -783,6 +788,8 @@ class StaticUpdater:
                     "upgrade_cost": level_data.get("UpgradeCost", 0),
                     "required_lab_level": level_data.get("LaboratoryLevel"),
                     "required_townhall": level_data.get("UpgradeLevelByTH") or self.lab_to_townhall[level_data.get("LaboratoryLevel")],
+                    "strength_weight": level_data.get("StrengthWeight", 0),
+
                 }
                 hold_data["levels"].append(new_level_data)
 
@@ -837,6 +844,8 @@ class StaticUpdater:
 
                     "required_townhall": level_data.get("RequiredTownHallLevel"),
                     "required_hero_tavern_level": level_data.get("RequiredHeroTavernLevel"),
+
+                    "strength_weight": level_data.get("StrengthWeight", 0),
                 }
                 hold_data["levels"].append(new_level_data)
 
@@ -889,6 +898,8 @@ class StaticUpdater:
                     "upgrade_cost": level_data.get("UpgradeCost", 0),
                     "required_pet_house_level": level_data.get("LaboratoryLevel"),
                     "required_townhall": self.pethouse_to_townhall[level_data.get("LaboratoryLevel")],
+
+                    "strength_weight": level_data.get("StrengthWeight", 0),
                 }
                 hold_data["levels"].append(new_level_data)
 
@@ -951,6 +962,7 @@ class StaticUpdater:
                     "heal_on_activation" : level_data.get("HealOnActivation", 0),
                     "required_blacksmith_level": level_data.get("RequiredBlacksmithLevel"),
                     "required_townhall": self.smithy_to_townhall[level_data.get("RequiredBlacksmithLevel")],
+                    "strength_weight": level_data.get("StrengthWeight", 0),
                     "upgrade_cost": {
                         "shiny_ore": shiny_ore,
                         "glowy_ore": glowy_ore,
@@ -1030,6 +1042,7 @@ class StaticUpdater:
                     "build_time": upgrade_time_seconds,
                     "required_townhall": level_data.get("TownHallLevel"),
                     "damage": level_data.get("Damage", 0),
+                    "strength_weight": level_data.get("StrengthWeight", 0),
                 })
 
             new_trap_data.append(hold_data)